\name{adg}
\docType{data}
\alias{adg}
\title{Average Daily Gain Data}
\description{Data example used here come from "SAS System for Mixed Models" by Little et. al. (1996) and it is an example data within SASmixed library. This dataset was collected to investigate average daily gains (ADG)
 of steers fed for 160 days. The treatments are four diets consisting
 of a base ration and three levels of a medicated feed additive added
 to the base ration. The objective of the experiment was to determine
 the optimal level of feed additive to maximize the average daily gain.
 The steers were housed in barns, the blocking factor, where each barn
 held four steers and the steers were individually fed. For the purposes of illustration we imposed missing values on the response values, following a missing completely at random mechanism.
}

\usage{data(adg)}
\format{A data frame containing 32 rows and 7 columns( response variables: first two columns which are average daily gain in grams, initial weight on the log scale ; predictors next 7 rows (a column of 1 and 3 treatment dummies) and subject indicator is last column)
}

\source{Littell, R.C., Milliken, G.A., Stroup, W.W., and Wolfinger, R.D. (1996). SAS System for Mixed Models, SAS Institute, Cary, NC.}
\keyword{datasets}

