\name{pain}
\docType{data}
\alias{pain}
\title{The longitudinal study on pain in children}
\description{This data consist of up to four observations on 64 children 8 to 10. The response is the length of time in seconds that the child can tolerate keeping  his or her arm in very cold water, a proxy measure of pain tolerance. Missing data exist due to absenteeism, broken arms, or other reasons. Two measures were taken during a first visit followed by two more measures during a second visit after a two-week gap.
}

\usage{data(pain)}
\format{A data frame containing 256 rows and 10 columns( response variables: first two rows; predictors next 7 rows and subject indicator is last column)
}

\source{Weiss, R.E. (2005) Modeling Longitudinal Data. New York, Springer.}
\keyword{datasets}

